added averageAccessesPerMonth method
this method figures out what the avarge usage each month based on how many years the log cointains. It will then make a array list to store the avarges for each month.  NOTE: the array list for the avarges are in int so it wont be exsact but it will be the avarge without the decimal.