extra credit
had to fix the LofFileCreator so that it made a new file with the proper ammounts of each code before I wasnt getting any 403s because my if statment was messed up. added a codeCountsPerMonth method so that you can see how many of each code appear on avarge a month for a year given in the parameter.